+++ CHANGELOG.md
@@ -70,6 +70,8 @@ Putting this stuff down here so things don't conflict as often.
 * Fixed invasions started by TShock not reporting size correctly and probably not working at all. (@hakusaro)
 * Removed `GetDataHandlers.TileKill` and replaced it with `GetDataHandlers.PlaceChest` as the packet originally designated as tile kill is now only used for chests. (@hakusaro)
 * Added `TSPlayer` to `GetDataHandlers.NPCHome`. (@hakusaro)
+* Added a `TSPlayer` to `GetDataHandlers.ChestItemChanged`. (@hakusaro)
+* Fixed chest item changes not triggering any range checks, tile checks, or correct chest checks. (@hakusaro)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.